+++ R/grex-datalist.R
@@ -6,3 +6,12 @@
 #' @format A character vector with 56,318 Ensembl IDs.
 #' @source \url{http://www.gtexportal.org}
 'gtexv6'
+
+#' Ensembl IDs from GTEx V6p Gene Read Count Data
+#'
+#' A dataset containing the Ensembl IDs from GTEx (V6p)
+#' gene read count data.
+#'
+#' @format A character vector with 56,238 Ensembl IDs.
+#' @source \url{http://www.gtexportal.org}
+'gtexv6p'