+++ R/grex-package.R
@@ -13,5 +13,4 @@
 #' @author Nan Xiao <\email{me@@nanx.me}>,
 #'         Gao Wang <\email{gaow@@uchicago.edu}>,
 #'         Lei Sun <\email{sunl@@uchicago.edu}>
-#'
 NULL