+++ NEWS.md
@@ -3,3 +3,11 @@
 ## NEW FEATURES
 
  * First version
+ 
+# CHANGES IN grex VERSION 1.4 (2016-11-20)
+
+## NEW FEATURES
+
+ * New GTEx V6p Ensembl ID data: `gtexv6p`
+ * Update reference database from the latest release version of `org.Hs.eg.db` (3.4.0)
+ * Add attribute "gene type" to the mapping table (Thanks: Prof. Matthew Stephens)